requirements/_base.txt
Django==1.5.2
South==0.8.2
django-autoslug==1.6.1
django-braces==1.2.2
